Implemented registering routing policies and added tests
